+++ tests/testthat/test_figures_taxo.R
@@ -10,21 +10,21 @@ data_fungi_2trees <-
 GP_archae <-
   subset_taxa(GlobalPatterns, GlobalPatterns@tax_table[, 1] == "Archaea")
 
-test_that("rotl_pq works with data_fungi dataset", {
-  skip_on_os("windows")
-  skip_on_cran()
-  library("rotl")
-  expect_s3_class(suppressWarnings(tr <-
-    rotl_pq(data_fungi, species_colnames = "Genus_species")), "phylo")
-  expect_s3_class(suppressWarnings(
-    rotl_pq(
-      data_fungi,
-      species_colnames = "Genus_species",
-      context_name = "Ascomycetes"
-    )
-  ), "phylo")
-  expect_silent(plot(tr))
-})
+# test_that("rotl_pq works with data_fungi dataset", {
+#   skip_on_os("windows")
+#   skip_on_cran()
+#   library("rotl")
+#   expect_s3_class(suppressWarnings(tr <-
+#                                      rotl_pq(data_fungi, species_colnames = "Genus_species")), "phylo")
+#   expect_s3_class(suppressWarnings(
+#     rotl_pq(
+#       data_fungi,
+#       species_colnames = "Genus_species",
+#       context_name = "Ascomycetes"
+#     )
+#   ), "phylo")
+#   expect_silent(plot(tr))
+# })
 
 # test_that("heat_tree_pq works with data_fungi dataset", {
 #   skip_on_cran()