+++ NEWS.md
@@ -1,10 +1,16 @@
-# MiscMetabar 0.10.3 (in development)
+# MiscMetabar 0.10.5 (in development)
+
+# MiscMetabar 0.10.4
+
+- Add functions `plot_refseq_pq()` and `plot_refseq_extremity_pq()` to plot the proportion of each nucleotide and the diversity of nucleotides from `@refseq` of a phyloseq object.
+
+# MiscMetabar 0.10.3 
 
 - Add params `type`, `na_remove` and `verbose` to `ggvenn_pq()`. The type = "nb_seq" allow to plot Venn diagram with the number of shared sequences instead of shared ASV. 
 - Add automatic report in json for the function `cutadapt_remove_primers()`.
 - Add param `verbose` to `track_wkflow()` and improve examples for `track_wkflow()` and `list_fastq_files`
 
-# MiscMetabar 0.10.2 (in development)
+# MiscMetabar 0.10.2
 
 - Improve code thanks to {lintr} package
 - Add option `return_file_path` to `cutadapt_remove_primers()` in order to facilitate targets pipeline